restructured project partially - still could be improved
