+++ Constants.py
@@ -1,0 +1,1 @@
+END_OF_QUEUE = None
