+++ Database/PostgresGameDAOFactory.py
@@ -1,0 +1,6 @@
+from Database.PostgresGameDAO import PostgresGameDAO
+from Database.PostgresConnectionFactory import PostgresConnectionFactory
+class PostgresGameDAOFactory:
+    @staticmethod
+    def createGameDAO():
+        return PostgresGameDAO(PostgresConnectionFactory)
